Display multiple panels per section (#120)
* Display multiple panels per section

We were overwriting the section each time we came across a new metric
within the section. Instead, append the new panel. Appending to a nil
slice will naturally create a new slice so this is safe to do.

* Add multiple metrics in group to model_metrics test

This test fails before the fix to putting groups together to catch any
regressions. In addition, upgrade the version of godeltaprof being used
in order to allow tests to be ran on a mac.

* Sort each section for consistent ordering

This both fixes an occassional failing test and will render the page in
a deterministic way.